+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add external change listener for SessionStorage and LocalStorage
+- Add documentation
 
 ## [1.0.2]