+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0]
+
 ### Added
 
 - Add external change listener for SessionStorage and LocalStorage
@@ -32,7 +34,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 First version
 
-[Unreleased]: https://github.com/MacFJA/svelte-persistent-store/compare/1.0.2...HEAD
+[Unreleased]: https://github.com/MacFJA/svelte-persistent-store/compare/1.1.0...HEAD
+[1.1.0]: https://github.com/MacFJA/svelte-persistent-store/compare/1.0.2...1.1.0
 [1.0.2]: https://github.com/MacFJA/svelte-persistent-store/compare/1.0.1...1.0.2
 [1.0.1]: https://github.com/MacFJA/svelte-persistent-store/compare/1.0.0...1.0.1
 [1.0.0]: https://github.com/MacFJA/svelte-persistent-store/releases/tag/1.0.0